WIP: Leader election.
The current limitation in the code is that even though we are
able to elect a leader for a particular term, that leader is not
able to establish its authority with the other servers before their
election timeout hits. This is adversely impacting liveness, we are
seeing re-election after re-election. This is because when a server
times out, it comes back with a bigger term number. It sends a RequestVote
RPC to the current leader with a bigger term number, and the leader steps
down and becomes follwer since it thinks that its state is outdated.