Resolving deadlock between AppendEntries RPC handler & electionTimeoutRoutine.
Send to channel rf.appendEntriesChannel was blocked since the
election goroutine was not able to reach the statement where
the channel is consumed. This was because AppendEntries RPC handler
was holding rf.mu.Lock() and at the same time the election goroutine
went into the election timeout clause and started waiting on the lock
to get released.